Eliminate bad security declarations through interface.
By using the ZCML class-require-permission directive, all direct or
indirect inherited interface method will be marked with the configured
permission on the class.
Usually we have no methods on content type interface, but in this
situation we are actually inheriting the "getFile" and "setFile"
methods, leading to an invalid security declaration for "setFile".

We should therefore no longer inherit the IFileContent interface in our
content type interface, but provide it seperately.