Introduce new permission "ftw.file: Edit date fields".
Protect the edit form tab (schemata) for the expiration date and effective date with a separate permission. This is useful if you want to grant the permission to edit these field to a role without the need to grant "manage content settings" which would allow the user to edit all the "advanced" fields.