Optimize disposition for most used Browsers
IE and Edge require a different disposition than Chrome, Firefox and Safari
so we need to differ them. I implemented that by checking in the user_agent
for some keywords. It is possible that the keywords do not match all the
possible user_agent cases. Since there isn't a finite number of possible
user_agent I think covering the most common cases is perfectly fine.