Dexterity migration: import GS profile even when not migrating
When a environment variable is set while running the AT to DX
migration upgrade step, the content migration itself is not executed.
This makes it possible for a policy package to control in what order
which types are migrated.

The existing upgrade step is changed so that it imports the generic
setup profile even when the skip-migration-env variable is activated.

Before all generic setup changes had to be copied to the policy
package, but we want to be able to only install the configuration
changes and migrate later without duplicating the settings.