Replace the upload utility with a separate package to prevent conflicts
Other packages need this utility as well and if we add it to all the
packages that need it we need tons of zcml conditions so that in the end
only one is registered. To prevent this we have extracted that utility
to a separate package and add it as a dependency in all the packages
that need it.